Bot restructure
I am restructuring the bot to relocate most functionality into extension files as cogs. This file will now handle running the bot and loading in extensions, and will hold just a few crucial commands.
The commands in this file are now set up so that most of the bot's functionality can be changed or updated on github, and remotely saved and loaded at runtime. via the bot's Discord command interface. This will allow for faster and easier updates and will lower the need for restarts.
I have also upgraded the bot to work with Discord Py v2. This will allow for the use of slash commands and a variety of other features.